GoogleSheetsOneDimensionSearcher
This searches in a row or a column for cells whose content matches
a Predicate.

When searching in a row, you get column indexes that match, while when
you search in a column, you get row indexes that match.

Matched indexes are mapped to something else calling the mapper function
provided.